% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{pep_hxb2JPT}
\alias{pep_hxb2JPT}
\title{Peptide collection for HXB2 and JPT clades}
\format{A \code{GRanges} object. One row per peptide.}
\usage{
data(pep_hxb2JPT)
}
\description{
Peptide collection for HXB2 and JPT clades
}

